Rename pod_manager.py to k8s_instance_manager.py for better comparison in CR.
